fix FilterDrawer tests
