+++ src/components/FilterMenu/FilterDrawerItem.js
@@ -29,7 +29,7 @@ function FilterDrawerItem({ filterItem, currentFilter, filterAction }) {
         </Grid>
       </Grid>
       <FormGroup className="filter-drawer-item__options">
-        {filterItem.options.map((option, index) => (
+        {filterItem?.options?.map((option, index) => (
           <FormControlLabel
             key={`filter-drawer-item-option-${option}`}
             componentsProps={{
